chore(otlp): split metrics pipeline module to multiple files
